+++ packages/webpack-config-dev-server/.babelrc
@@ -1,0 +1,5 @@
+{
+  "presets": [
+    "metalab"
+  ]
+}
