+++ packages/webpack-config-dev-server/README.md
@@ -1,0 +1,11 @@
+# webpack-config-dev-server
+
+Setup a development server for [webpack].
+
+```javascript
+import devServer from 'webpack-config-dev-server';
+
+partial(config, devServer);
+```
+
+[webpack]: https://webpack.github.io
