+++ packages/webpack-config-dev-server/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "webpack-config-entry",
+  "version": "0.1.0",
+  "author": "Izaak Schroeder <izaak.schroeder@gmail.com>",
+  "repository": "izaakschroeder/webpack-config-metalab",
+  "license": "CC0-1.0",
+  "main": "dist/dev-server.webpack.config.js",
+  "scripts": {
+    "prepublish": "./node_modules/.bin/babel -d dist src"
+  },
+  "peerDependencies": {
+    "webpack": "*",
+    "webpack-udev-server": "*"
+  },
+  "devDependencies": {
+    "babel-cli": "^6.0.0",
+    "babel-preset-metalab": "^0.1.1"
+  }
+}
